NEEDED location provider. Doing useEffect hook on every reload was inefficient for loading location. It works smoothly now
